implement layer offset + refactor
