+++ example/example-isometric.tmx
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<map version="1.5" tiledversion="1.7.2" orientation="isometric" renderorder="right-down" width="20" height="20" tilewidth="111" tileheight="64" infinite="0" backgroundcolor="#176baa" nextlayerid="4" nextobjectid="1">
+<map version="1.5" tiledversion="1.7.2" orientation="isometric" renderorder="right-down" width="20" height="20" tilewidth="111" tileheight="64" infinite="0" backgroundcolor="#176baa" nextlayerid="5" nextobjectid="1">
  <tileset firstgid="1" source="cubes.tsx"/>
  <layer id="1" name="Ground" width="20" height="20">
   <data encoding="csv">
@@ -50,6 +50,30 @@
 116,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,116,
 116,0,0,0,0,0,0,0,0,0,100,100,100,100,100,0,0,0,0,116,
 116,116,116,116,116,116,116,116,116,116,116,116,116,116,116,116,116,116,116,116
+</data>
+ </layer>
+ <layer id="4" name="PlayerLevelCollider" width="20" height="20">
+  <data encoding="csv">
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,116,116,116,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,116,116,116,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,116,116,116,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
 </data>
  </layer>
  <layer id="3" name="Above 1" width="20" height="20">
