+++ src/index.ts
@@ -8,3 +8,11 @@ export * from './tiled-object';
 export * from './tiled-tileset';
 export * from './tiled-object-component';
 export * from './tiled-layer-component';
+
+// raw data structures
+export * from './raw-tiled-layer';
+export * from './raw-tiled-map';
+export * from './raw-tiled-object';
+export * from './raw-tiled-text';
+export * from './raw-tiled-tileset';
+export * from './raw-tileset-tile';
