+++ src/tiled-types.ts
@@ -3,86 +3,6 @@ import {
 } from 'excalibur';
 import { TiledObject } from '.';
 
-/**
- * Tiled Map Interface
- *
- * Represents the interface for the Tiled exported data structure (JSON). Used
- * when loading resources via Resource loader.
- */
-export interface RawTiledMap {
-   type: 'map';
-   ex: ExcaliburData;
-   version: number;
-   
-   width: number;
-   /**
-    * Number of tile rows
-    */
-   height: number;
-   /**
-    * Length of the side of a hex tile in pixels (hexagonal maps only)
-    */
-   hexsidelength: number;
-   /**
-    * Map grid height
-    */
-   tileheight: number;
-   /**
-    * Map grid width
-    */
-   tilewidth: number;
-
-   /**
-    *	Hex-formatted color (#RRGGBB or #AARRGGBB) (optional)
-    */
-   backgroundcolor?: string;
-   /**
-    * The compression level to use for tile layer data (defaults to -1, which means to use the algorithm default)
-    */
-   compressionlevel: number;
-   /**
-    * Whether the map has infinite dimensions
-    */
-   infinite: boolean;
-   /**
-    *	Auto-increments for each layer
-    */
-   nextlayerid: number;
-   /**
-    * Auto-increments for each placed object
-    */
-   nextobjectid: number;
-   
-   /**
-    * Map orientation (orthogonal, isometric, staggered or hexagonal)
-    */
-   orientation: 'orthogonal' | 'isometric' | 'staggered' | 'hexagonal';
-
-   layers: RawTiledLayer[];
-   properties: TiledProperty[];
-   tilesets: RawTiledTileset[];
-
-   /**
-    * Render order: right-down (the default), right-up, left-down or left-up (currently only supported for orthogonal maps)
-    */
-   renderorder: 'right-down' | 'right-up' | 'left-down' | 'left-up';
-
-   /**
-    * x or y (staggered / hexagonal maps only)
-    */
-   staggeraxis: 'x' | 'y';
-   /**
-    * odd or even (staggered / hexagonal maps only)
-    */
-   staggerindex: 'odd' | 'even';
-
-   /**
-    * The Tiled version used to save the file
-    */
-   tiledversion: string;
-
-}
-
 export interface ExcaliburData {
    camera?: ExcaliburCamera;
    colliders?: ExcaliburCollider[];
@@ -125,107 +45,6 @@ export interface TiledProperty<T = unknown> {
 export type TiledEncoding = 'csv' | 'base64';
 export type TiledCompression = 'zlib' | 'gzip' | 'zstd';
 
-export interface RawTiledLayer {
-   /**
-    * Incremental ID - unique across all layers
-    */
-   id: number;
-   /**
-    * Image used by this layer. imagelayer only.
-    */
-   image: string;
-   /**
-    * Array of unsigned int (GIDs) or base64-encoded data. tilelayer only.
-    */
-   data: number[] | string;
-   /**
-    * 	Array of chunks (optional). tilelayer only.
-    */
-   chunks: TiledChunk[];
-   /**
-    * Column count. Same as map width for fixed-size maps.
-    */
-   width: number;
-   /**
-    * Row count. Same as map height for fixed-size maps.
-    */
-   height: number;
-   /**
-    * Name assigned to this layer
-    */
-   name: string;
-   /**
-    * From [0, 1]
-    */
-   opacity: number;
-   properties: TiledProperty[];
-   /**
-    * csv (default) or base64. tilelayer only.
-    */
-   encoding: TiledEncoding;
-   /**
-    * zlib, gzip, zstd (since Tiled 1.3) or empty (default). tilelayer only.
-    */
-   compression?: TiledCompression;
-
-   /**
-    * Type of layer (tilelayer, objectgroup)
-    */
-   type: 'tilelayer' | 'objectgroup' | 'imagelayer' | 'group';
-   /**
-    * Whether layer is shown or hidden in editor
-    */
-   visible: boolean;
-
-   /**
-    * Horizontal layer offset in tiles. Always 0.
-    */
-   x: number;
-   /**
-    * Vertical layer offset in tiles. Always 0.
-    */
-   y: number;
-
-   /**
-    * Layer order in the original Tiled source
-    */
-   order: number;
-   /**
-    * Horizontal layer offset in pixels (default: 0)
-    */
-   offsetx: number;
-   /**
-    * Vertical layer offset in pixels (default: 0)
-    */
-   offsety: number;
-   /**
-    * X coordinate where layer content starts (for infinite maps)
-    */
-   startx: number;
-   /**
-    * Y coordinate where layer content starts (for infinite maps)
-    */
-   starty: number;
-
-   /**
-    * Hex-formatted color (#RRGGBB or #AARRGGBB) that is multiplied with any graphics drawn by this layer or any child layers (optional).
-    */
-   tintcolor: string;
-   /**
-    * Hex-formatted color (#RRGGBB) (optional). imagelayer only.
-    */
-   transparentcolor: string;
-
-   /**
-    * topdown (default) or index. objectgroup only.
-    */
-   draworder: 'topdown' | 'index' | 'objectgroup';
-   /**
-    * Array of objects. objectgroup only.
-    */
-   objects: RawTiledObject[];
-}
-
 export interface TiledChunk {
    /**
     * Array of unsigned int (GIDs) or base64-encoded data
@@ -247,201 +66,6 @@ export interface TiledChunk {
     * Y coordinate in tiles
     */
    y: number;
-}
-
-export interface RawTiledObject {
-   id: number;
-
-   /**
-    * Tile object id
-    */
-   gid: number;
-   /**
-    * Used to mark an object as a point
-    */
-   point: boolean;
-   height: number;
-   name: string;
-   properties: TiledProperty[];
-   /**
-    * Angle in degrees clockwise
-    */
-   rotation: number;
-   type: string;
-   visible: boolean;
-   width: number;
-   /**
-    * X coordinate in pixels
-    */
-   x: number;
-   /**
-    * Y coordinate in pixels
-    */
-   y: number;
-
-   /**
-    * Reference to a template file, in case object is a template instance
-    */
-   template: string;
-
-   /**
-    *	Only used for text objects
-    */
-   text: RawTiledText;
-
-   /**
-    * Whether or not object is an ellipse
-    */
-   ellipse: boolean;
-
-   /**
-    * Polygon points
-    */
-   polygon: TiledPoint[];
-
-   /**
-    * Polyline points
-    */
-   polyline: TiledPoint[];
-}
-
-export interface RawTiledText {
-   text: string;
-   /**
-    * Whether to use a bold font (default: false)
-    */
-   bold: boolean;
-   /**
-    * Hex-formatted color (#RRGGBB or #AARRGGBB) (default: #000000)
-    */
-   color: string;
-   /**
-    * Font family (default: sans-serif)
-    */
-   fontfamily: string;
-   /**
-    * Horizontal alignment (center, right, justify or left (default))
-    */
-   halign: 'center' | 'right' | 'justify' | 'left';
-   /**
-    * Whether to use an italic font (default: false)
-    */
-   italic: boolean;
-   /**
-    * Whether to use kerning when placing characters (default: true)
-    */
-   kerning: boolean;
-   /**
-    * Pixel size of font (default: 16)
-    */
-   pixelsize: number;
-   /**
-    * Whether to strike out the text (default: false)
-    */
-   strikeout: boolean;
-   /**
-    * Whether to underline the text (default: false)
-    */
-   underline: boolean;
-   /**
-    * Vertical alignment (center, bottom or top (default))
-    */
-   valign: 'center' | 'bottom' | 'top';
-   /**
-    * Whether the text is wrapped within the object bounds (default: false)
-    */
-   wrap: boolean;
-}
-
-export interface RawTiledTileset {
-   type: 'tileset';
-   /**
-    * The JSON format version
-    */
-   version: number;
-
-   /**
-    * GID corresponding to the first tile in the set
-    */
-   firstgid: number;
-   /**
-    * Image used for tiles in this set
-    */
-   image: string;
-   /**
-    * Height of source image in pixels
-    */
-   imageheight: number;
-   /**
-    * Width of source image in pixels
-    */
-   imagewidth: number;
-   /**
-    * (optional)
-    */
-   grid: TiledGrid;
-   /**
-    * Buffer between image edge and first tile (pixels)
-    */
-   margin: number;
-   /**
-    * Alignment to use for tile objects (unspecified (default), topleft, top, topright, left, center, right, bottomleft, bottom or bottomright) (since 1.4)
-    */
-   objectalignment: 'unspecified' | 'topleft' | 'top' | 'topright' | 'left' | 'center' | 'right' | 'bottomleft' | 'bottom' | 'bottomright'
-   /**
-    * Refers to external tileset file (should be JSON)
-    */
-   source: string;
-   /**
-    * Spacing between adjacent tiles in image (pixels)
-    */
-   spacing: number;
-
-   columns: number;
-   rows: number;
-   /**
-    * Maximum height of tiles in this set
-    */
-   tileheight: number;
-   tilewidth: number;
-
-   /**
-    * Array of Tiles (optional)
-    */
-   tiles: RawTilesetTile[];
-
-   name: string;
-   properties: TiledProperty[];
-   /**
-    * The number of tiles in this tileset
-    */
-   tilecount: number;
-   /**
-    * Optional
-    */
-   tileoffset: TiledTileOffset;
-
-   /**
-    * The Tiled version used to save the file
-    */
-   tiledversion: string;
-   /**
-    * Hex-formatted color (#RRGGBB or #AARRGGBB) (optional)
-    */
-   backgroundcolor: string;
-   /**
-    * Hex-formatted color (#RRGGBB) (optional)
-    */
-   transparentcolor: string;
-   /**
-    * Array of Terrains (optional)
-    */
-   terrains: TiledMapTerrain[];
-
-   /**
-    * Array of Wang sets (since 1.1.5)
-    */
-   wangsets: TiledWangSet[]
 }
 
 export interface TiledTileOffset {
@@ -529,20 +153,6 @@ export interface TiledGrid {
     */
    height: number;
 }
-
-export interface RawTilesetTile {
-   id: number;
-   type: string;
-   image: string;
-   imageheight: number;
-   imagewidth: number;
-   animation: TiledFrame[];
-   properites: TiledProperty[];
-   terrain: number[];
-   objectgroup: RawTiledLayer;
-   probability: number;
-}
-
 
 export interface TiledFrame {
    /**
